3: Update React app for build process & vice-versa
