+++ .browserslistrc
@@ -1,3 +1,4 @@
 # Supported Browsers
 
-last 2 versions
+# i.e. "defaults" as of 2019
+> 0.5%, last 2 versions, Firefox ESR, not dead
