+++ public/index.html
@@ -5,11 +5,11 @@
 		<meta charset="utf-8">
 		<meta http-equiv="X-UA-Compatible" content="IE=edge">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<link href="index.css" rel="stylesheet">
-		<script type="module" src="index.mjs"></script>
-		<script nomodule defer src="index.js"></script>
+		<link rel="stylesheet" href="index.css">
+		<script type="module" src="native/index.mjs"></script>
+		<script nomodule defer src="native/index.js"></script>
 	</head>
-	<body>
+	<body id="page--index">
 		<header>
 			<h1><abbr title="What You See is What You Get">WYSIWYG</abbr> Card Maker</h1>
 			<p>Fake card. Fake game. Real fun.</p>
