--- src/index.js
@@ -1,3 +1,0 @@
-import init from './native/index.js';
-
-init();
