+++ src/react/components/ShapeInput.js
@@ -1,11 +1,11 @@
-// NOTE: If processing, use `import`
-// import React, { useState, useEffect } from 'react';
-// import PropTypes from 'prop-types';
-// import CustomTypes from './custom-types.js';
-const { useState, useEffect } = window.React, React = window.React;
-const PropTypes = window.PropTypes;
-const CustomTypes = window.CustomTypes;
-const { Select } = window;
+import React from 'react';
+import PropTypes from 'prop-types';
+
+// Components
+import Select from './Select.js';
+
+// Data
+import CustomTypes from '../services/custom-types.js';
 
 /**
  * Called when input/field value changes
@@ -30,13 +30,13 @@ const { Select } = window;
  */
 function ShapeInput( props ) {
 	const { id, label, desc, shapes: options, value: initialValue, labelClassName, outputClassName, onChange, ...jsxAttrs } = props;
-	const [ value, setValue ] = useState( initialValue );
+	const [ value, setValue ] = React.useState( initialValue );
 
 	// FAQ: We can manage change internally and externally
 	function handleChange( e ) {
 		setValue( e.target.value );
 	}
-	useEffect(() => {
+	React.useEffect(() => {
 		if ( onChange ) onChange( value );
 	}, [ value ]);
 
@@ -63,3 +63,5 @@ ShapeInput.propTypes = {
 	outputClassName: PropTypes.string,
 	onChange: PropTypes.func,
 }
+
+export default ShapeInput;
